Deallocate the objects
This resets the RCP instance to null, which causes the underlying object's
reference count to be decreased, possibly deallocating it.

Without this patch, the RCP gets deallocated when it runs out of scope, which
sometimes happens late or never (in the Cython generated code, the RCP sits
inside a struct). This patch sets it to null, so we deallocate the C++ object
when the Python object gets deallocated, which is the right time to do it.