Add support for 32 and 16 bit numpy floats
This commit adds support for 32 and 16 bit floats numpy as input types.
Since it doesn't look like there is a native symengine type for working
with single (or half) precision floats outside of using MPFR (which
seemed like the wrong approach here, it might make sense for np.float128
in a follow up though) this implicitly casts the np.float16 and
np.float32 types to doubles and just behaves as float/np.float64.

Fixes #351